+++ src/features/project/hooks/useCreateProjectMutation.ts
@@ -1,0 +1,33 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { projectApi } from '@/features/project/api/projectApi';
+import type { Project } from '@/features/project/types/projectTypes';
+import { generateId } from '@/shared/utils/idUtils';
+
+// 프로젝트 생성
+export const useCreateProjectMutation = () => {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: projectApi.createProject,
+    onMutate: async () => {
+      await queryClient.cancelQueries({ queryKey: ['projects'] });
+      const previousProjects = queryClient.getQueryData(['projects']);
+
+      const newTempProject: Project = {
+        projectId: `temp-${generateId()}`,
+        name: `새로운 프로젝트`,
+        defaultReviewerCount: 2,
+        role: 'admin',
+      };
+
+      queryClient.setQueryData(['projects'], (old: Project[]) => {
+        return [...(old || []), newTempProject];
+      });
+
+      return { previousProjects };
+    },
+
+    onSettled: () => {
+      queryClient.invalidateQueries({ queryKey: ['projects'] });
+    },
+  });
+};
