+++ src/features/project/hooks/useDeleteProjectMutation.ts
@@ -1,0 +1,11 @@
+import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { projectApi } from '@/features/project/api/projectApi';
+
+// 프로젝트 삭제
+export const useDeleteProjectMutation = () => {
+  const queryClient = useQueryClient();
+  return useMutation({
+    mutationFn: projectApi.deleteProject,
+    onSuccess: () => queryClient.invalidateQueries({ queryKey: ['projects'] }),
+  });
+};
